analysis/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "analytics-engine"
version = "0.1.0"
description = "Analytics engine components for streaming training and prediction"
readme = "README.md"
requires-python = ">=3.10"
authors = [ { name = "Aakash Mhatre" } ]

dependencies = [
  "accelerate==1.11.0",
  "aiohttp==3.13.2",
  "email-validator==2.3.0",
  "evaluate==0.4.6",
  "fastapi==0.100.0",
  "httptools==0.7.1",
  "itsdangerous==2.2.0",
  "orjson==3.11.4",
  "pipdeptree==2.30.0",
  "pydantic-extra-types==2.10.6",
  "python-multipart==0.0.20",
  "scikit-learn==1.7.2",
  "setuptools==58.1.0",
  "transformers==4.57.1",
  "ujson==5.11.0",
  "uvicorn==0.38.0",
  "uvloop==0.22.1",
  "watchfiles==1.1.1",
  "websockets==15.0.1",
]

[tool.setuptools.packages.find]
where = ["analytics-engine"]